% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{all_counts}
\alias{all_counts}
\title{literature_object after pvalue testing
Result of the test_score method of the pubscore class}
\format{A list}
\usage{
data(all_counts)
}
\description{
literature_object after pvalue testing
Result of the test_score method of the pubscore class
}
\keyword{datasets,}
\keyword{literature}
\keyword{pubmed,}
\keyword{test,}
