% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubscore.R
\docType{methods}
\name{AllCounts}
\alias{AllCounts}
\alias{AllCounts,PubScore-method}
\title{Retrieve the all_counts attribute}
\usage{
AllCounts(pub, ...)

\S4method{AllCounts}{PubScore}(pub)
}
\arguments{
\item{pub}{Object of class \code{PubScore}}
}
\value{
A dataframe containing the counts table for all genes.
}
\description{
Retrieve the all_counts attribute
}
\examples{
# Create a new pubscore object
pub <- pubscore(genes = c('cd4','cd8'),terms_of_interest = c('blabla','immunity'))
plot(networkViz(pub))
}
