% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubscore.R
\name{.getSimulation_test}
\alias{.getSimulation_test}
\title{Auxiliary function for the test method}
\usage{
.getSimulation_test(pub, ambiguous = c(), n_simulations)
}
\description{
Auxiliary function for the test method
}
