% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscore.gene.R
\docType{data}
\name{gene}
\alias{gene}
\title{iscore.gene}
\format{An object of class \code{character} of length 1.}
\usage{
gene
}
\arguments{
\item{terms_of_interest}{A list of terms of interest related to the topic you want to find the relevance for}

\item{gene}{The gene which you want to calculate the iscore for}
}
\description{
Calculates the importance score for a given gene.
The importance score is the total counts of articles in the pubmed
database that are a result for that gene AND each term of a list
}
\keyword{datasets}
