% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{gene2pubmed_db}
\alias{gene2pubmed_db}
\title{human genes on gene2pubmed_db}
\format{A list}
\usage{
data(gene2pubmed_db)
}
\description{
human genes on gene2pubmed_db
}
\keyword{datasets,}
\keyword{literature}
\keyword{pubmed,}
\keyword{test,}
