% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubscore.R
\docType{methods}
\name{get_all_counts}
\alias{get_all_counts}
\alias{get_all_counts,PubScore-method}
\title{Retrieve the all_counts attribute}
\usage{
get_all_counts(pub, ...)

\S4method{get_all_counts}{PubScore}(pub)
}
\arguments{
\item{pub}{Object of class \code{PubScore}}
}
\value{
A dataframe containing the counts table for all genes.
}
\description{
Retrieve the all_counts attribute
}
\examples{
# Create a new pubscore object
pub <- pubscore(genes = c('cd4','cd8'),terms_of_interest = c('blabla','immunity'))
plot(networkViz(pub))
}
