% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_iscore.R
\name{get_iscore}
\alias{get_iscore}
\title{get_iscore}
\usage{
get_iscore(gene, terms_of_interest, is.list = F, max.score = 500,
  verbose = T)
}
\arguments{
\item{gene}{The gene which you want to calculate the iscore for, or a vector with multiple genes}

\item{terms_of_interest}{A list of terms of interest related to the topic you want to find the relevance for}

\item{is.list}{If you are searching a single gene or a list of genes. Defaults to false (single gene)}

\item{max.score}{A cutoff for maximum numbers of search. Useful for avoiding outlier relations
weighting too much. Defaults to 500.}
}
\description{
Calculates the importance score for a given gene.
The importance score is the total counts of articles in the pubmed
database that are a result for that gene AND each term of a list
}
\examples{
gene <- 'CD4'
terms_of_interest <- c("CD4 T cell", "CD14+ Monocyte", "B cell", "CD8 T cell",
                      "FCGR3A+ Monocyte", "NK cell", "Dendritic cell", "Megakaryocyte", 'immunity')
hm <- get_iscore(gene, terms_of_interest, max.score = 500)
get_iscore(gene, terms_of_interest, max.score = Inf)
}
