% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_literature_score.R
\name{get_literature_score}
\alias{get_literature_score}
\title{get_literature_score}
\usage{
get_literature_score(gene, terms_of_interest, is.list = F,
  max_score = Inf, wait_time = 0)
}
\arguments{
\item{gene}{The gene which you want to calculate the literature_score for, or a vector with multiple genes}

\item{terms_of_interest}{A list of terms of interest related to the topic you want to find the relevance for}

\item{is.list}{If you are searching a single gene or a list of genes. Defaults to false (single gene)}

\item{max_score}{A cutoff for maximum numbers of search. Useful for avoiding outlier relations
weighting too much. Defaults to Inf.}

\item{wait_time}{How long should be the interval between two requests to the ENTREZ database when it fails.
Defaults to 0. In seconds.}
}
\description{
Calculates the importance score for a given gene.
The importance score is the total counts of articles in the pubmed
database that are a result for that gene AND each term of a list
}
\examples{
gene <- 'CD4'
terms_of_interest <-  c("CD4 T cell","CD14+ Monocyte", "B cell", "CD8 T cell",
"FCGR3A+ Monocyte", "NK cell",  "Dendritic cell",    "Megakaryocyte",    'immunity'  )
get_literature_score(gene, terms_of_interest, 500)
get_literature_score(gene, terms_of_interest, max_score = Inf)
}
