% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubscore.R
\docType{methods}
\name{getScore}
\alias{getScore}
\alias{getScore,PubScore-method}
\title{Retrieve the literature_score attribute}
\usage{
getScore(pub)

\S4method{getScore}{PubScore}(pub)
}
\arguments{
\item{pub}{Object of class \code{PubScore}}
}
\value{
A "numeric" with the literature score for this gene x term combination
}
\description{
Retrieve the literature_score attribute
}
\examples{
# Create a new pubscore object
pub <- pubscore(genes = c('cd4','cd8'),terms_of_interest = c('blabla','immunity'))
plot(networkViz(pub))
}
