% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubscore.R
\docType{methods}
\name{heatmapViz}
\alias{heatmapViz}
\alias{heatmapViz,PubScore-method}
\title{SECTION 2 - METHODS TO RETRIEVE ATTRIBUTES
Retrieve the heatmap attribute}
\usage{
heatmapViz(pub)

\S4method{heatmapViz}{PubScore}(pub)
}
\arguments{
\item{pub}{Object of class \code{PubScore}}
}
\value{
A "gg" object, from ggplot2, containing a heatmap from the counts table.
}
\description{
SECTION 2 - METHODS TO RETRIEVE ATTRIBUTES
Retrieve the heatmap attribute
}
\examples{
#Create a new pubscore object
pub <- pubscore(genes = c('cd4','cd8'),terms_of_interest = c('blabla','immunity'))
plot(heatmapViz(pub))
}
