% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{hgcn_entrez_reference}
\alias{hgcn_entrez_reference}
\title{hgcn_entrez_reference
biomaRt results (october 2019) for entrezids and gene symbols.
processed.}
\format{A list}
\usage{
data(hgcn_entrez_reference)
}
\description{
hgcn_entrez_reference
biomaRt results (october 2019) for entrezids and gene symbols.
processed.
}
\keyword{datasets,}
\keyword{literature}
\keyword{pubmed,}
\keyword{test,}
