% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iscore_plot.R
\name{iscore_plot}
\alias{iscore_plot}
\title{iscore_plot}
\usage{
iscore_plot(plot_counts, return_ggplot = F, is_plotly = F)
}
\arguments{
\item{plot_counts}{The dataframe returned as the second object ($counts) in the
list output of get_iscore function}

\item{return_ggplot}{If TRUE, returns a ggplot2 object instead of plotting. Defaults to FALSE.}

\item{is_plotly}{If TRUE, a interactive plot is plotted in the place o static ggplot. Defaults to FALSE.}
}
\description{
Plots a non-clusterized heatmap of the article counts for the combination of
gene list and list of terms
}
\examples{
gene <- 'CD4'
terms_of_interest <- c("CD4 T cell", "CD14+ Monocyte", "B cell", "CD8 T cell",
                      "FCGR3A+ Monocyte", "NK cell", "Dendritic cell", "Megakaryocyte", 'immunity')
iscore_list <- get_iscore(gene, terms_of_interest, max.score = 500)
iscore_plot(iscore_list$counts)
}
