% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{literature_object}
\alias{literature_object}
\title{literature_object after pvalue testing
Result of the test literature score function}
\format{A list}
\usage{
data(literature_object)
}
\description{
literature_object after pvalue testing
Result of the test literature score function
}
\examples{
data(literature_object)
print(literature_object$p_value)
retest_literature_score(literature_object, remove_ambiguous = TRUE, nsim = 10000 )
}
\keyword{datasets,}
\keyword{literature}
\keyword{pubmed,}
\keyword{test,}
