% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_literature_score_graph.R
\name{plot_literature_graph}
\alias{plot_literature_graph}
\title{#'plot_literature_graph}
\usage{
plot_literature_graph(plot_counts, name, color = "#B30000FF",
  max_number_of_labels = 10)
}
\arguments{
\item{plot_counts}{The dataframe returned from the get_literature_score function}

\item{name}{The name of the plot.}

\item{color}{The color of the plot. Defaults to a shade of red ("#B30000FF").}

\item{max_number_of_labels}{The max number of gene labels to show.  Defaults to 10.}
}
\value{
A ploty/ggplot2 object is either returned or directly plotted
}
\description{
Plot a graph inspired in CEMiTool's graphs
}
\examples{
 gene <- c('CD4','CD14', "AIF1", "ACVR1", "CDY2A")
 terms_of_interest <- c("CD4 T cell", "CD14+ Monocyte")
 literature_counts <- get_literature_score(gene, terms_of_interest)
  pl <- plot_literature_graph(literature_counts, name = 'test')
  pl
}
