% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_literature_score.R
\name{plot_literature_score}
\alias{plot_literature_score}
\title{plot_literature_score}
\usage{
plot_literature_score(plot_counts, return_ggplot = F, is_plotly = F)
}
\arguments{
\item{plot_counts}{The dataframe returned as the second object ($counts) in the
list output of get_literature_score function}

\item{return_ggplot}{If TRUE, returns a ggplot2 object instead of plotting. Defaults to FALSE.}

\item{is_plotly}{If TRUE, a interactive plot is plotted in the place o static ggplot. Defaults to FALSE.}
}
\description{
Plots a non-clusterized heatmap of the article counts for the combination of
gene list and list of terms
}
\examples{
gene <- 'CD4'
terms_of_interest <- c("CD4 T cell", "CD14+ Monocyte", "B cell", "CD8 T cell",
                      "FCGR3A+ Monocyte", "NK cell", "Dendritic cell", "Megakaryocyte", 'immunity')
literature_list <- get_literature_score(gene, terms_of_interest, max_score = 500)
plot_literature_score(literature_score_list$counts)
}
