% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retest_literature_score.R
\name{retest_literature_score}
\alias{retest_literature_score}
\title{retest_literature_score}
\usage{
retest_literature_score(literature_object,
  new_ambigous_term_list = FALSE, remove_ambiguous = FALSE,
  nsim = 1e+05)
}
\arguments{
\item{literature_object}{The object returned by the test_literature_score function}

\item{new_ambigous_term_list}{A vector containing a list of genes deemed ambiguous by the user. 
If left blank, a default list is used.}

\item{remove_ambiguous}{If TRUE, ambiguously named genes (such as "MARCH") will be removed. 
Defaults to TRUE.}

\item{nsim}{The number of simulations to run. Defaults to 100000.}
}
\value{
A dataframe with the literature scores.
}
\description{
Regets a pvalue, by simulation, for a literature score.
}
\examples{

genes <- c('CD14')
terms_of_interest <-
  c(
    "CD4 T cell",
    "CD14+ Monocyte",
    "B cell",
    "CD8 T cell",
    "FCGR3A+ Monocyte",
    "NK cell",
    "Dendritic cell",
    "Megakaryocyte",
    'immunity'
  )
literature_object <- PubScore::get_literature_score(genes, terms_of_interest)
total_genes <- c('CD14','notagenemock1','notagenemock2','notagenemock3', "PROC", "IMPACT")
literature_object <- test_literature_score(literature_object, total_genes, remove_ambiguous = FALSE)
literature_object <- retest_literature_score(literature_object, remove_ambiguous = TRUE)

}
