% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubscore.R
\docType{methods}
\name{set_all_counts<-}
\alias{set_all_counts<-}
\alias{set_all_counts<-,PubScore-method}
\title{Set the all_counts attribute}
\usage{
set_all_counts(pub) <- value

\S4method{set_all_counts}{PubScore}(pub) <- value
}
\arguments{
\item{pub}{Object of class \code{PubScore}}

\item{value}{The table with all gene x term article counts from the "test_score" method.}
}
\value{
A dataframe containing the counts table for all genes.
}
\description{
Set the all_counts attribute
}
\examples{
terms_of_interest <- c('Dengue')
pub <- pubscore(terms_of_interest = terms_of_interest, genes = c("CD4", "CD8", "CD14") )
print(getScore(pub))
data("all_counts")
set_all_counts(pub) <- all_counts
}
