% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_literature_score.R
\name{test_literature_score}
\alias{test_literature_score}
\title{get_literature_score}
\usage{
test_literature_score(literature_object, total_genes,
  show_progress = TRUE, verbose = FALSE)
}
\arguments{
\item{literature_object}{The object returned by the get_literature_score function}

\item{total_genes}{A list of all the possible genes in your study. 
Usually all the names in the rows of an "exprs" object.}

\item{show_progress}{If TRUE, a progress bar is displayed. Defaults to True.}

\item{verbose}{If TRUE, will display the index of the search occuring. Defaults to false.}
}
\value{
A dataframe with the literature scores.
}
\description{
Calculates the importance score for a given gene.
The importance score is the total counts of articles in the pubmed
database that are a result for that gene AND each term of a list
}
\examples{
   genes <- c('CD4','CD8')
   terms_of_interest <-
     c(
       "CD4 T cell",
       "CD14+ Monocyte",
       "B cell",
       "CD8 T cell",
       "FCGR3A+ Monocyte",
       "NK cell",
       "Dendritic cell",
       "Megakaryocyte",
       'immunity'
     )
 literature_object <- PubScore::get_literature_score(gene, terms_of_interest, 500)
 total_genes <- c('CD4', 'CD8','CD7','ASDASD','ADSFASD','ADSFASDS', 'asdas')
 literature_object <- test_literature_score(literature_object, total_genes)
}
