% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubscore.R
\docType{methods}
\name{test_score}
\alias{test_score}
\alias{test_score,PubScore-method}
\title{Get and test the literature enrichment score}
\usage{
test_score(pub, total_genes, show_progress = TRUE,
  remove_ambiguous = TRUE, verbose = FALSE, nsim = 1e+05)

\S4method{test_score}{PubScore}(pub, total_genes, show_progress = TRUE,
  remove_ambiguous = TRUE, verbose = FALSE, nsim = 1e+05)
}
\arguments{
\item{pub}{Object of class \code{PubScore}}

\item{total_genes}{A list of all the possible genes in your study. 
Usually all the names in the rows of an "exprs" object.}

\item{show_progress}{If TRUE, a progress bar is displayed. Defaults to True.}

\item{remove_ambiguous}{If TRUE, ambiguously named genes (such as "MARCH") will be removed. Defaults to TRUE.}

\item{verbose}{If TRUE, will display the index of the search occuring. Defaults to false.}

\item{nsim}{The number of simulations to run. Defaults to 100000.}
}
\value{
A "gg" object, from ggplot2, containing a network from the counts table.
}
\description{
Get and test the literature enrichment score
}
\examples{
Create a new pubscore object
pub <- pubscore(genes = c('cd4','cd8'),terms_of_interest = c('blabla','immunity'))
pub <- test_score(pub, total_genes = c('notagene1', 'notagene2', 'cd4', 'cd8'),remove_ambiguous = T)
}
