% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{all_counts}
\alias{all_counts}
\title{all_counts}
\format{A list}
\usage{
data(all_counts)
}
\description{
A dataframe with all counts for total genes in the Dengue dataset
and the term Dengue. Outcome of the test_score method of the pubscore class
}
\keyword{datasets,}
\keyword{literature}
\keyword{pubmed,}
\keyword{test,}
