% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{all_counts}
\alias{all_counts}
\title{all_counts}
\format{
An object of class \code{data.frame}
}
\usage{
data(all_counts)
}
\description{
A dataframe with all pubmed counts  for the genes in the Dengue dataset in relation 
to the term "Dengue".
}
\details{
Outcome of the test_score method of the pubscore class.
As this function may take a long time, this dataset speeds up the vignette.

Contains:
3 columns: 
#tax_id: The reference ID for the taxon. All are 9606 (humans).
GeneID: The Entrez ID code for a given gene.
PubMedID: A PubMed ID for a paper that mentions the gene in the "Gene ID" column.

1335548 rows: gene-paper associations in the gene2pubmed database.
}
\keyword{datasets,}
\keyword{literature}
\keyword{pubmed,}
\keyword{test,}
