% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_literature_score.R
\name{get_literature_score}
\alias{get_literature_score}
\title{get_literature_score}
\usage{
get_literature_score(genes, terms_of_interest, gene2pubmed = FALSE,
  wait_time = 0, show_progress = TRUE, verbose = FALSE)
}
\arguments{
\item{genes}{A vector with multiple genes.}

\item{terms_of_interest}{A list of terms of interest related to the topic you want to find the relevance for}

\item{gene2pubmed}{Boolean defining if gene2pubmed db is going to be used.
If used, the vector of genes has to be of HUMAN genes in the hgcn_symbol format.}

\item{wait_time}{How long should be the interval between two requests to the ENTREZ database when it fails.
Defaults to 0. In seconds.}

\item{show_progress}{If TRUE, a progress bar is displayed. Defaults to TRUE.}

\item{verbose}{If TRUE, will display the index of the search occuring. Defaults to FALSE.}

\item{return_all}{Only to be used with gene2pubmed! Boolean defining if the 
all_counts table is going to be returned here. Usually it is generated by the
test_score function.}
}
\value{
A dataframe with the literature scores.
}
\description{
Calculates the importance score for a given gene.
The importance score is the total counts of articles in the pubmed
database that are a result for that gene AND each term of a list
}
\examples{
genes <- c('CD8A', 'CD4')
terms_of_interest <-
  c(
    "CD4 T cell",
    "CD14+ Monocyte",
    "B cell",
    "CD8 T cell",
    "FCGR3A+ Monocyte",
    "NK cell",
    "Dendritic cell",
    "Megakaryocyte",
    'immunity'
  )
get_literature_score(genes, terms_of_interest)
}
