% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubscore.R
\docType{methods}
\name{initialize,PubScore-method}
\alias{initialize,PubScore-method}
\title{initialize}
\usage{
\S4method{initialize}{PubScore}(.Object, genes, terms_of_interest,
  gene2pubmed = FALSE)
}
\arguments{
\item{.Object}{The object of signature PubScore that is foinf to be created}

\item{genes}{The genes to which you want to calculate and visualize the literature score.}

\item{terms_of_interest}{A list of terms of interest related to the topic you want to find the relevance.}

\item{gene2pubmed}{Boolean defining if gene2pubmed db is going to be used.}
}
\value{
A object of the PubScore class
}
\description{
initialize
}
