% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_literature_score.R
\name{plot_literature_score}
\alias{plot_literature_score}
\title{plot_literature_score}
\usage{
plot_literature_score(plot_counts, return_ggplot = FALSE,
  is_plotly = FALSE)
}
\arguments{
\item{plot_counts}{The dataframe returned from the get_literature_score function}

\item{return_ggplot}{If TRUE, returns a ggplot2 object instead of plotting. Defaults to FALSE.}

\item{is_plotly}{If TRUE, a interactive plot is plotted in the place o static ggplot. Defaults to FALSE.}
}
\value{
A ggplot2 object is either returned or directly plotted
}
\description{
Plots a non-clusterized heatmap of the article counts for the combination of
gene list and list of terms
NOTE: the object has to be exactly the one returned by get_literature_score.R .
Otherwise ggplot2 will not be able to identify the correct columns.
}
\examples{
  gene <- c('CD4','CD14', "AIF1", "ACVR1", "CDY2A")
  terms_of_interest <- c("CD4 T cell", "CD14+ Monocyte", "B cell", "CD8 T cell",
                          "FCGR3A+ Monocyte", "NK cell", "Dendritic cell", "Megakaryocyte", 'immunity')
  literature_counts <- get_literature_score(gene, terms_of_interest)
  P <-plot_literature_score(literature_counts, return_ggplot = TRUE)
  plot(P)
}
