% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubscore.R
\docType{class}
\name{PubScore-class}
\alias{PubScore-class}
\title{An S4 class to represent PubScore results}
\description{
An S4 class to represent PubScore results
}
\section{Slots}{

\describe{
\item{\code{terms_of_interest}}{A list of terms of interest related to the topic you want to find the relevance.}

\item{\code{genes}}{The genes to which you want to calculate and visualize the literature score.}

\item{\code{date}}{The date when the object was initialized. PubScore counts will likely increase with time.}

\item{\code{counts}}{A data frame with the counts retrieved on PubMed}

\item{\code{network}}{A visualization of the results found in a network}

\item{\code{heatmap}}{A visualization of the results found in a heatmap}
}}

