% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_literature_score.R
\name{test_literature_score}
\alias{test_literature_score}
\title{test_literature_score}
\usage{
test_literature_score(literature_object, total_genes,
  show_progress = TRUE, remove_ambiguous = TRUE, verbose = FALSE)
}
\arguments{
\item{literature_object}{The object returned by the get_literature_score function}

\item{total_genes}{A list of all the possible genes in your study. 
Usually all the names in the rows of an "exprs" object.}

\item{show_progress}{If TRUE, a progress bar is displayed. Defaults to True.}

\item{remove_ambiguous}{If TRUE, ambiguously named genes (such as "MARCH") will be removed. Defaults to TRUE.}

\item{verbose}{If TRUE, will display the index of the search occuring. Defaults to false.}
}
\value{
A dataframe with the literature scores.
}
\description{
gets a pvalue, by simulation, for a literature score.
}
\examples{
     genes <- c('CD14')
     terms_of_interest <-
       c(
         "CD4 T cell",
         "CD14+ Monocyte",
         "B cell",
         "CD8 T cell",
         "FCGR3A+ Monocyte",
         "NK cell",
         "Dendritic cell",
         "Megakaryocyte",
         'immunity'
       )
   literature_object <- PubScore::get_literature_score(genes, terms_of_interest)
   total_genes <- c('CD14', 'notagenemock1','notagenemock2','notagenemock3', "PROC", "IMPACT")
   literature_object <- test_literature_score(literature_object, total_genes)
 
}
